Plugin UT: etcd start of day with two ports but no existing etcd data
