+++ calico/openstack/test/lib.py
@@ -18,8 +18,10 @@ openstack.test.lib
 
 Common code for Neutron driver UT.
 """
+from eventlet.support import greenlets as greenlet
 import mock
 import sys
+import traceback
 
 sys.modules['neutron'] = m_neutron = mock.Mock()
 sys.modules['neutron.common'] = m_neutron.common
@@ -29,6 +31,26 @@ sys.modules['neutron.plugins'] = m_neutron.plugins
 sys.modules['neutron.plugins.ml2'] = m_neutron.plugins.ml2
 sys.modules['neutron.plugins.ml2.drivers'] = m_neutron.plugins.ml2.drivers
 sys.modules['etcd'] = m_etcd = mock.Mock()
+
+port1 = {'binding:vif_type': 'tap',
+         'binding:host_id': 'felix-host-1',
+         'id': 'DEADBEEF-1234-5678',
+         'device_owner': 'compute:nova',
+         'fixed_ips': [{'subnet_id': '10.65.0/24',
+                        'ip_address': '10.65.0.2'}],
+         'mac_address': '00:11:22:33:44:55',
+         'admin_state_up': True,
+         'security_groups': ['SGID-default']}
+
+port2 = {'binding:vif_type': 'tap',
+         'binding:host_id': 'felix-host-1',
+         'id': 'FACEBEEF-1234-5678',
+         'device_owner': 'compute:nova',
+         'fixed_ips': [{'subnet_id': '10.65.0/24',
+                        'ip_address': '10.65.0.3'}],
+         'mac_address': '00:11:22:33:44:66',
+         'admin_state_up': True,
+         'security_groups': ['SGID-default']}
 
 
 # Define a stub class, that we will use as the base class for
@@ -52,6 +74,12 @@ class Lib(object):
     osdb_ports = []
 
     def setUp(self):
+        # Announce the current test case.
+        print "\nTEST CASE: %s" % self.id()
+
+        # Hook logging.
+        self.setUp_logging()
+
         # If an arg mismatch occurs, we want to see the complete diff of it.
         self.maxDiff = 1000
 
@@ -64,6 +92,9 @@ class Lib(object):
 
         # Arrange what the DB's get_ports will return.
         self.db.get_ports.side_effect = lambda *args: self.osdb_ports
+
+        # Arrange DB's get_subnet call.
+        self.db.get_subnet.return_value = {'gateway_ip': '10.65.0.1'}
 
         # Arrange what the DB's get_security_groups query will return (the
         # default SG).
@@ -100,3 +131,33 @@ class Lib(object):
         # Prep a null response to the following
         # _get_port_security_group_bindings call.
         self.db._get_port_security_group_bindings.return_value = []
+
+    def setUp_logging(self):
+        """Setup to intercept and display logging by the code under test.
+        """
+        # Print logs to stdout.
+        def log_info(msg):
+            print "       INFO %s" % msg
+            return None
+        def log_debug(msg):
+            print "       DEBUG %s" % msg
+            return None
+        def log_warn(msg):
+            print "       WARN %s" % msg
+            return None
+        def log_error(msg):
+            print "       ERROR %s" % msg
+            return None
+        def log_exception(msg):
+            print "       EXCEPTION %s" % msg
+            if sys.exc_type is not greenlet.GreenletExit:
+                traceback.print_exc()
+            return None
+
+        # Hook logging.
+        mech_calico.LOG = mock.Mock()
+        mech_calico.LOG.info.side_effect = log_info
+        mech_calico.LOG.debug.side_effect = log_debug
+        mech_calico.LOG.warn.side_effect = log_warn
+        mech_calico.LOG.error.side_effect = log_error
+        mech_calico.LOG.exception.side_effect = log_exception
